pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "viva-datacreator"
version = "0.0.3"
description = "An open-source tool for creating image segmentation datasets from videos using SAM 2"
readme = "README.md"
authors = [
  { name = "Miguel S. Soriano-García, Diego Mercado-Ravell, Israel Becerra, Julio De La Torre-Vanegas", email = "m16soriano@gmail.com" }
]
requires-python = ">=3.12"
license = { text = "MIT" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "pyyaml>=6.0.3",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    # "sam-2 @ git+https://github.com/facebookresearch/sam2.git",
    "sam2",
    "markupsafe<3.0",
    "opencv-python>=4.11.0.86",
    "matplotlib>=3.10.7",
    "sahi>=0.11.36",
    "pandas>=2.3.3",
    "ipython>=9.7.0",
    "ultralytics>=8.3.228",
    "tqdm>=4.67.1",
    "deep-sort-realtime>=1.3.2",
    "setuptools<71",
    "typer>=0.20.0",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[project.scripts]
viva-creator = "vivadatacreator.guimain:main"
download-checkpoints = "vivadatacreator.download_checkpoints:main"
install = "vivadatacreator.install:main"

[project.optional-dependencies]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.17",
    "mkdocstrings[python]>=0.26.1",
]

[tool.setuptools.packages.find]
include = ["vivadatacreator*"]

[tool.setuptools.package-data]
vivadatacreator = [
    "class_dict.csv",
    "GUI/**/*"
]
